sil-scafold-project/backend/requirements.txt
Django>=4.2
djangorestframework>=3.14
djangorestframework-simplejwt>=5.2.0
psycopg2-binary>=2.9
django-mptt>=0.13
celery>=5.3
redis>=4.5
python-dotenv>=1.0
requests>=2.31
python-jose>=3.3.0
africastalking>=1.2.5
django-nose>=1.4.7
drf-yasg>=1.21.0
graphene-django>=3.0.0
django-redis>=5.3.0
django-cors-headers>=4.1.0
coverage>=7.0.0
pytest>=7.0.0
pytest-django>=4.5.0
pytest-cov>=4.0.0
pytest-mock>=3.10.0
